+++ docs/examples/events/on_decorator02.py
@@ -4,7 +4,7 @@ from textual.widgets import Button
 
 
 class OnDecoratorApp(App):
-    CSS_PATH = "on_decorator.css"
+    CSS_PATH = "on_decorator.tcss"
 
     def compose(self) -> ComposeResult:
         """Three buttons."""