+++ tests/test_tabbed_content.py
@@ -475,6 +475,38 @@ async def test_disabling_via_tabbed_content():
         assert app.query_one(Tabs).active == "tab-1"
 
 
+async def test_disabling_via_tab_pane():
+    class TabbedApp(App[None]):
+        def compose(self) -> ComposeResult:
+            with TabbedContent():
+                yield Label("tab-1")
+                yield Label("tab-2")
+
+        def on_mount(self) -> None:
+            self.query_one("TabPane#tab-2").disabled = True
+
+    app = TabbedApp()
+    async with app.run_test() as pilot:
+        await pilot.pause()
+        await pilot.click("Tab#tab-2")
+        assert app.query_one(Tabs).active == "tab-1"
+
+
+async def test_creating_disabled_tab():
+    class TabbedApp(App[None]):
+        def compose(self) -> ComposeResult:
+            with TabbedContent():
+                with TabPane("first"):
+                    yield Label("hello")
+                with TabPane("second", disabled=True):
+                    yield Label("world")
+
+    app = TabbedApp()
+    async with app.run_test() as pilot:
+        await pilot.click("Tab#tab-2")
+        assert app.query_one(Tabs).active == "tab-1"
+
+
 async def test_navigation_around_disabled_tabs():
     class TabbedApp(App[None]):
         def compose(self) -> ComposeResult:
@@ -539,6 +571,29 @@ async def test_reenabling_via_tabbed_content():
 
     app = TabbedApp()
     async with app.run_test() as pilot:
+        await pilot.click("Tab#tab-2")
+        assert app.query_one(Tabs).active == "tab-1"
+        app.reenable()
+        await pilot.click("Tab#tab-2")
+        assert app.query_one(Tabs).active == "tab-2"
+
+
+async def test_reenabling_via_tab_pane():
+    class TabbedApp(App[None]):
+        def compose(self) -> ComposeResult:
+            with TabbedContent():
+                yield Label("tab-1")
+                yield Label("tab-2")
+
+        def on_mount(self) -> None:
+            self.query_one("TabPane#tab-2").disabled = True
+
+        def reenable(self) -> None:
+            self.query_one("TabPane#tab-2").disabled = False
+
+    app = TabbedApp()
+    async with app.run_test() as pilot:
+        await pilot.pause()
         await pilot.click("Tab#tab-2")
         assert app.query_one(Tabs).active == "tab-1"
         app.reenable()
@@ -696,3 +751,55 @@ async def test_showing_first_tab_activates_tab(tab_id: str):
         tabbed_content.show_tab(tab_id)
         await pilot.pause()
         assert tabbed_content.active == tab_id
+
+
+async def test_disabling_nested_tabs():
+    """Regression test for https://github.com/Textualize/textual/issues/3145."""
+
+    class TabbedApp(App):
+        def compose(self) -> ComposeResult:
+            with TabbedContent(id="tabbed-content"):
+                with TabPane("Tab Pane 1"):
+                    yield Label("foo")
+                with TabPane("Tab Pane 2"):
+                    yield Label("bar")
+                with TabPane("Tab Pane 3"):
+                    with TabbedContent():
+                        with TabPane("Inner Pane 1"):
+                            yield Label("fizz")
+                        with TabPane("Inner Pane 2"):
+                            yield Label("bang")
+
+    app = TabbedApp()
+    async with app.run_test() as pilot:
+        tabber = app.query_one("#tabbed-content", expect_type=TabbedContent)
+        tabber.disable_tab("tab-1")
+        await pilot.pause()
+        tabber.enable_tab("tab-1")
+        await pilot.pause()
+
+
+async def test_hiding_nested_tabs():
+    """Regression test for https://github.com/Textualize/textual/issues/3145."""
+
+    class TabbedApp(App):
+        def compose(self) -> ComposeResult:
+            with TabbedContent(id="tabbed-content"):
+                with TabPane("Tab Pane 1"):
+                    yield Label("foo")
+                with TabPane("Tab Pane 2"):
+                    yield Label("bar")
+                with TabPane("Tab Pane 3"):
+                    with TabbedContent():
+                        with TabPane("Inner Pane 1"):
+                            yield Label("fizz")
+                        with TabPane("Inner Pane 2"):
+                            yield Label("bang")
+
+    app = TabbedApp()
+    async with app.run_test() as pilot:
+        tabber = app.query_one("#tabbed-content", expect_type=TabbedContent)
+        tabber.hide_tab("tab-1")
+        await pilot.pause()
+        tabber.show_tab("tab-1")
+        await pilot.pause()